refactor: move remaining files to new packages
