+++ app/src/main/java/org/feature/fox/coffee_counter/ui/profile/ProfileView.kt
@@ -1,10 +1,11 @@
-package org.feature.fox.coffee_counter.ui
+package org.feature.fox.coffee_counter.ui.profile
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import org.feature.fox.coffee_counter.R
+import org.feature.fox.coffee_counter.ui.common.MoneyAppBar
 
 @Preview(showSystemUi = true)
 @Composable
