+++ app/src/main/java/org/feature/fox/coffee_counter/ui/user/EditUserView.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterialApi::class)
 
-package org.feature.fox.coffee_counter.ui
+package org.feature.fox.coffee_counter.ui.user
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
